fix bug with df_rate (num_items would not be lost during import and filter any more)

+++ src/artm/core/dictionary.h
@@ -77,8 +77,8 @@ class Dictionary {
   const DictionaryEntry* entry(const Token& token) const;
   const DictionaryEntry* entry(int index) const;
 
-  size_t size() const { return entries_.size(); }
-  size_t num_items() const { return num_items_in_collection_; }
+  int size() const { return entries_.size(); }
+  int num_items() const { return num_items_in_collection_; }
   const std::string& name() const { return name_; }
   bool has_valid_cooc_state() const;
   int64_t ByteSize() const;
